requirements.txt
web3==6.15.1
asyncio-mqtt
requests>=2.31.0
python-dotenv>=1.0.0
psutil>=5.9.5
python-telegram-bot>=20.0